requirements.txt
blinkstick~=1.2.0
cryptography~=44.0.1
flask~=3.1.0
flask-login~=0.6.3
flask-sqlalchemy~=3.1.1
mysql-connector-python~=8.0.0
python-dateutil~=2.9.0
python-dotenv~=1.0.0
requests~=2.31.0
SQLAlchemy~=2.0.0
pandas~=2.2.0
google-auth~=2.35.0
google-auth-oauthlib~=1.2.0
google-auth-httplib2~=0.2.0
google-api-python-client~=2.149.0